Support config file in argument in kali-mirror-images
So that we can use this script to mirror daily images. Right now, the
script to mirror daily images is a copy/paste of this one, the only
difference is that it hardcodes different values for TO, RSYNC_PATH and
log file names.

With this change, we can now use the same script to mirror daily images,
we just need to make it use a different config file.

For loading commandline arguments, we respect the same logic as ftpsync,
ie. give preference to arguments passed to the script, and otherwise
check if SSH_ORIGINAL_COMMAND is set. Cf. bin/ftpsync, from line 354.

To avoid blindly opening a user-supplied file, the script is rather
restrictive in what it accepts as a configuration file. Actually the
argument is not even a conf file, it's just a component that is used
to form the name of the config file -> etc/mirror-COMPONENT.conf