Make the log filename depend on COMPONENT in mirror-kali-images
Now that we can have several instances of the script running at once,
with different config files, we need to make sure that each instance
with log to a different log file.

The most obvious is just to re-use the COMPONENT in the log file name,
which is already what's done anyway with the default component
kali-images.

Additionally, this commit embeds the 'log/' directory in the variable,
it seems less error prone this way, We already embed the 'etc/'
directory in the CONFIG_FILE variable above, so let's follow this style
consistently.