Make sure log dir exists before running rsync
The 'log' directory is created by ftpsync. If ever, on a new deployment,
mirror-kali-images runs before ftpsync, I suppose that the log dir
doesn't exist and it will fail.

Note that the documentation at [1] explicitly ask users to run a first
sync manually, but it asks users to run *rsync* (rather than *ftpsync*).
It doesn't seem to ask to run ftpsync manually.

So let's be on the safe side and create the log directory before trying
to use it.

[1] https://www.kali.org/docs/community/setting-up-a-kali-linux-mirror/