Wrap rsync command with a lock in mirror-kali-images
Otherwise several instances of this script might run in parallel. For
example, if the system administrator runs it by hand, while Cron runs it
at the same time. Or, who knows, if the networks is very slow and the
script is scheduled to run at too short interval, so that it doesn't
have time to complete when a new run is scheduled.

To prevent that, let's use a lock. To keep it simple, we just do a
non-blocking lock that fails if the lock is already taken.

We use the 'locks/' directory, that is created by ftpsync (cf.
bin/common). We use the config file as the name for the lock, as it's
already guaranteed to be unique, and it gives us one lock per config
file, which is what we want.

Note that all the scripts in archvsync have been written to be portable
to multiple Unix having rsync available. Kali mirrors can run on *BSD or
other kind of non-standard systems, therefore flock is optional, and if
it's not found, we do without (no lock). We even go as far as making
sure that the arguments (-x and -n) are supported by Busybox' flock.
Just in case.

If the lock can't be taken, we make sure to return a identifiable exit
code that is *not* among the range of documented rsync exit codes.

Also, note that we have to be careful not to touch the log files if we
can't take the lock. Indeed, if we can't take it, it's because another
instance is running and writing to these log files, so that last thing
we want to do is to erase it with some kind of error message from
flock). That's why we don't do "flock rsync ... > $LOGFILE", instead we
use the rather complicated "sub shell construct", which was not invented
here, but is documented in flock(1).